34: Migrate to Peer library.
This (almost) finally fixes my problem.
By doing the migration, I found that I was using `transact` incorrectly:
I passed tx-data as a top-level arg instead of putting it in a map
under the `:tx-data` key.

This approach is fine for Peer library but the APIs are different.
Now it's indeed transacting new data.