Lesson 49 + 50: Clear notifications + tests.
This implements DELETE on `/conversations/<conversation-id>` such that
it marks all messages in given conversation as read-by the account id set in the Authorization header.
This effectively means "clearing notifications" for the account.

This DELETE operation is special (doesn't really remove any entity) so we cannot reuse `crud/delete`.
Instead, we implement custom operation that first calls `find-unread-messages`,
then adds the request account id to `:message/read-by` attribute's value.

`find-unread-messages` is unlike all the other operations in route handlers,
in that it has to query the db directly, instead of delegating to _interceptors_.
This is done by exposing `interceptors/query!` as a public function.