54. play with cognito via aws-api.
This is the basic setup adding the dependencies
and constructing the client.

We construct the client by providing proper credentials profile
and also setting the region explicitly to "us-east-1",
otherwise it doesn't work with my aws config (probably defaulting to "eu-west-1").

We enable spec validation with `(aws/validate-requests cognito true)`

We finally try to call `aws/invoke` and get the expected error message:

```
  (aws/invoke cognito
              {:op :SignUp
               ;; look here to find the client information: https://us-east-1.console.aws.amazon.com/cognito/v2/idp/user-pools/us-east-1_naCTUkfCg/app-integration/clients/60c14ln0t3saa1jg2e0q13oj2v?region=us-east-1
               :request {:ClientId "60c14ln0t3saa1jg2e0q13oj2v"
                         :Username"jumar" ; this is aws username
                         :Password "xxx"}})
  ;; => {:__type "NotAuthorizedException",
  ;;     :message "Client 60c14ln0t3saa1jg2e0q13oj2v is configured for secret but secret was not received",
  ;;     :cognitect.anomalies/category :cognitect.anomalies/incorrect}
```

We'll fix this in the next lesson providing proper :SecretHash.